backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-socketio[asyncio_client]==5.11.1
aiohttp>=3.9.0
celery==5.3.6
redis==5.0.1
anthropic>=0.39.0
openai>=1.50.0
alembic==1.13.1
python-multipart==0.0.9
